9043/refactor: extract certificate calculation logic into CertificateCalculatorService
